+++ app/components/header.jsx
@@ -1,10 +1,8 @@
-import { Link, useLocation } from '@remix-run/react';
+  import { Link } from '@remix-run/react';
 import logo from '../../public/img/logo.svg'
+import Navegation from './navegation';
 
 function Header() {
-
-  const location = useLocation()
-  console.log(location)
 
   return (
     <header className="header">
@@ -12,27 +10,9 @@ function Header() {
         <Link to='/'>
           <img className='logo' src={logo} alt='imagen logo' />
         </Link>
-        <nav className="navegacion">
-          <Link
-            to='/'
-            className={location.pathname === '/' ? 'active' : '' }
-          >Inicio</Link>
 
-          <Link
-            to='/us'
-            className={location.pathname === '/us' ? 'active' : '' }
-          >Us</Link>
+        <Navegation />
 
-          <Link
-            to='/store'
-            className={location.pathname === '/store' ? 'active' : '' }
-          >Store</Link>
-
-          <Link
-            to='/blog'
-            className={location.pathname === '/blog' ? 'active' : '' }
-          >Blog</Link>
-        </nav>
       </div>
     </header>
   )
