+++ app/root.jsx
@@ -1,6 +1,7 @@
 import { Meta, Links, Outlet, Scripts, LiveReload } from '@remix-run/react';
 import styles from './styles/index.css'
 import Header from './components/header';
+import Footer from './components/footer';
 
 export function meta() {
   return [
@@ -55,7 +56,7 @@ function Document({ children }) {
       <body>
         <Header />  
         {children}
-      
+        <Footer />
         <Scripts />
         <LiveReload />
       </body>
